% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.R
\name{my_theme}
\alias{my_theme}
\title{My theme for ggplot2 publication ready graphs}
\usage{
my_theme(base_size = 14, base_family = "arial")
}
\description{
Publication ready GGplot2 graphs
}
\section{Functions}{
\itemize{
\item \code{my_theme}: color palette for my graphs
}}

\examples{
\dontrun{
# do not run this
g2 <- g + my_theme()
creates g2, a ggplot2 object with my personal theme.
}
}
\author{
Salil Deo
}
