% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_na.R
\name{input_na}
\alias{input_na}
\title{Convert values to NA}
\usage{
input_na(x, y)
}
\arguments{
\item{x}{your input col, needs to be in the format df$x}

\item{y}{if more than 1 value, then use c(a,b,c) format}
}
\description{
Convert any specified values to NA in list, column
}
\details{
A function to convert to NA in column in dataframe
}
\examples{
\dontrun{
x <- c(1,2,-9)
z <- input_na(x,-9)
}
\dontrun{
this can also be used for more than 1 value
x <- c(1,2,3,4,5,-9,-99)
z <- input_na(x, c(-9,-99))
}
}
\keyword{missing}
