% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_cmp.R
\name{proc_cmp}
\alias{proc_cmp}
\title{A function to create competing risk curve
    for 1 group with user specified labels}
\usage{
proc_cmp(c, times, color)
}
\arguments{
\item{c}{a competing risk object with cuminc function}

\item{times}{time to label survival values provide as list}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\description{
Competing risk curve with 1 group
}
\seealso{
\code{\link[survival]{survfit}}

\code{\link[cmprsk]{cuminc}}
}
\author{
Salil Deo
}
