% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_desc.R
\name{proc_desc}
\alias{proc_desc}
\title{Summary statistics with histogram}
\usage{
proc_desc(x)
}
\arguments{
\item{x}{list,vector,column of a dataframe}
}
\description{
summary statistics with histogram of a variable.
}
\examples{
\dontrun{
x = c(1,2,3,4,5,6,7,8,9,10)
proc_desc(x)
}
}
