% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_km_label.R
\name{proc_km_label}
\alias{proc_km_label}
\title{A function to create personalized KM curve for 1 group}
\usage{
proc_km_label(s, xlab, ylab, color, times)
}
\arguments{
\item{s}{survfit object}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{color}{string providing color label}

\item{times}{time for labelling survival values provide as list}
}
\description{
Label survival values for KM curve
}
\examples{
\dontrun{
# do not run this
# needs my_theme already in .env
library(survival)
s = survfit(Surv(time, status) ~ 1, data = lung)
figure <- proc_kmcurve2(s = s, xlab = "follow-up",
ylab = "proportion surviving",
color = "blue", times = c(300,500))
gets an object figure is a ggplot2 object; it can be further modified if needed.
}
}
\seealso{
\code{\link[survival]{survfit}}
}
\author{
Salil Deo
}
