% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_km_label2.R
\name{proc_km_label2}
\alias{proc_km_label2}
\title{A function to create personalized KM curve for 2 or more groups}
\usage{
proc_km_label2(s, xlab, ylab, times)
}
\arguments{
\item{s}{survfit object}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{times}{time to label survival values provide as list}
}
\description{
Label survival values for KM curve
}
\examples{
\dontrun{
# do not run this
library(survival)
s = survfit(Surv(time, status) ~ sex, data = lung)
figure <- proc_kmcurve2(s = s, xlab = "follow-up",
ylab = "proportion surviving", times = c(200,500))
gets an object figure is a ggplot2 object; it can be further modified if needed.
}
}
\seealso{
\code{\link[survival]{survfit}}
}
\author{
Salil Deo
}
