% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_kmcurve1.R
\name{proc_kmcurve1}
\alias{proc_kmcurve1}
\title{A function to create personalized KM curve for 1 group}
\usage{
proc_kmcurve1(s, xlab, ylab, color)
}
\arguments{
\item{s}{survfit object}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{color}{string providing color label}
}
\description{
Provides KM curve for 1 group with CI band
}
\examples{
\dontrun{
# do not run this
library(survival)
s = survfit(Surv(time, status) ~ 1, data = lung)
figure <- proc_kmcurve2(s = s, xlab = "follow-up", ylab = "proportion surviving",
color = "blue")
gets an object figure is a ggplot2 object; it can be further modified if needed.
}
}
\seealso{
\code{\link[survival]{survfit}}
}
\author{
Salil Deo
}
