% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_kmcurve2.R
\name{proc_kmcurve2}
\alias{proc_kmcurve2}
\title{A function to create personalized KM curve for 2 or more groups}
\usage{
proc_kmcurve2(s, xlab, ylab)
}
\arguments{
\item{s}{survfit object}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\description{
Provides KM curve for 2/more groups with CI bands
}
\examples{
\dontrun{
# do not run this
library(survival)
s = survfit(Surv(time, status) ~ sex, data = lung)
figure <- proc_kmcurve2(s = s, xlab = "follow-up", ylab = "proportion surviving")
gets an object figure is a ggplot2 object; it can be further modified if needed.
}
}
\seealso{
\code{\link[survival]{survfit}}
}
\author{
Salil Deo
}
