% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_tab.R
\name{proc_tab}
\alias{proc_tab}
\title{A function to provide descriptive statistics.}
\usage{
proc_tab(x)
}
\arguments{
\item{x}{column in a dataframe}
}
\description{
Provides descriptive statistics as SPSS format.
}
\examples{
\dontrun{
# do not run this
x = c(1,2,3,4,5,6,7,8,9,10)
proc_tab(x)
}
}
\seealso{
\code{\link[gmodels]{CrossTable}}
}
